Drop the library search hack for running the tests
As this is a rather old feedstock, it relied on a hack on macOS and
Linux to ensure libraries were loaded correctly when running the tests
as `conda-build` had not worked it's prefix magic to fix the paths in
binaries to load them from the correct location. The workaround was to
set the library fallback paths for loading these libraries to search in
the correct location when nothing else worked. However SIP on macOS
10.11 made this no longer a viable solution. To fix this issue, simply
set the `rpath` correctly on macOS and Linux when building libraries.
This way the libraries will already have the right location to load
linked dependencies from and there is no need for the hack. Of course
`conda-build` will still do binary prefix substitution, which is fine as
well.